% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE1.R
\name{calculate_EDGE1}
\alias{calculate_EDGE1}
\title{Calculate EDGE1 index}
\usage{
calculate_EDGE1(
  tree,
  table,
  species.col = "species",
  RLcat.col = "RLcat",
  sort.list = TRUE,
  ...
)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{species.col}{column storing species names. If not specified, a column named "species" will be searched for.}

\item{RLcat.col}{column storing Red List assessments names. If not specified, a column named "RLcat" will be searched for.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}
}
\value{
a data frame with the input table and the EDGE metric (sensu  Isaac et al., 2007).
}
\description{
Calculate EDGE1 index
}
\references{
Isaac, N.J., Turvey, S.T., Collen, B., Waterman, C. & Baillie, J.E. (2007)
Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2: e296.
\url{https://doi.org/10.1371/journal.pone.0000296}
}
