% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE1.R
\name{calculate_EDGE1_multiphylo}
\alias{calculate_EDGE1_multiphylo}
\title{Calculate EDGE1 index for a multiphylo object}
\usage{
calculate_EDGE1_multiphylo(
  multitree,
  table,
  species.col = "species",
  RLcat.col = "RLcat",
  sort.list = FALSE,
  summarise = TRUE,
  parallelize = FALSE,
  n.cores = NULL,
  ...
)
}
\arguments{
\item{multitree}{a `multiPhylo` object storing multiple trees}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{species.col}{column storing species names. If not specified, a column named "species" will be searched for.}

\item{RLcat.col}{column storing Red List assessments names. If not specified, a column named "RLcat" will be searched for.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{summarise}{Logical. If FALSE, a list of dataframes outputted from `calculate_EDGE1` function will be returned,
one for each tree stored in the 'multitree' `multiPhylo` object. If FALSE, results will be summarised into a single table,
with mean, standard deviation, median and inter-quartile ranges for ED and EDGE metrics across trees.}

\item{parallelize}{Logical. If TRUE, several CPU cores (defined in `n.cores`) will be used.}

\item{n.cores}{Integer. Number of cores to use simultaneoulsly. Default is the number of available ones minus one.}
}
\value{
If `summarise` parameter is set to TRUE (default), a data frame with ED and EDGE averaged values. If not summarised,
a list storing results of `calculate_EDGE1` for each tree in the 'multitree' object.
}
\description{
Calculate EDGE1 index for a multiphylo object
}
\references{
Isaac, N.J., Turvey, S.T., Collen, B., Waterman, C. & Baillie, J.E. (2007)
Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE 2: e296.
\url{https://doi.org/10.1371/journal.pone.0000296}
}
