% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE2}
\alias{calculate_EDGE2}
\title{EDGE2 calculating function.}
\usage{
calculate_EDGE2(
  tree,
  table,
  species.col = "species",
  RLcat.col = "RLcat",
  ext.prob = "Isaac",
  verbose = TRUE,
  sort.list = FALSE,
  return.all = FALSE,
  seed = NULL
)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{species.col}{column storing species names. If not specified, a column named "species" will be searched for.}

\item{RLcat.col}{column storing Red List assessments names. If not specified, a column named "RLcat" will be searched for.}

\item{ext.prob}{Extinction probability associated to category, based on Isaac et al. (2007) or Mooers et al. (2008).}

\item{verbose}{Logical. Should progress be printed or not.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}
}
\value{
A list containing 3 slots if `return.all` is TRUE. If not, only the first item (EDGE list) will be returned.
\itemize{
 \item{list: }{will output specific TBL (terminal branch length), pext (sampled probability of extinction), ED (evolutionary distinctiveness), and EDGE (EDGE index).}
 \item{tree: }{the input tree.}
 \item{ePDloss: }{complete PD (phylogenetic diversity) and ePDloss (expected PD loss). Units are as in the input tree  (generally Million years).}
}
}
\description{
EDGE2 calculating function.
}
\references{
Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C., Baillie, J.E.M. (2007).
Mammals on the EDGE: Conservation Priorities Based on Threat and Phylogeny. PloS one, 2(3), e296.
\url{https://doi.org/10.1371/journal.pone.0000296}

Mooers, A.Ø., Faith, D. P., & Maddison, W. P. (2008).
 Converting endangered species categories to probabilities of extinction for phylogenetic
 conservation prioritization. PloS one, 3(11), e3700.
\url{https://doi.org/10.1371/journal.pone.0003700}

Gumbs, R., Gray, C. L., Böhm, M., Burfield, I. J., Couchman, O. R., Faith, D. P.,
Forest, F., Hoffmann, M., #' Isaac, N. J. B., Jetz, W., Mace, G. M., Mooers, A. O., Safi, K.,
Scott, O., Steel, M., Tucker, C. M., Pearse, W. D., Owen, N. R. & Rosindell, J. (2023).
The EDGE2 protocol: Advancing the prioritisation of Evolutionarily Distinct and Globally Endangered species for
practical conservation action. PLoS Biology, 21(2), e3001991.
\url{https://doi.org/10.1371/journal.pbio.3001991}
}
\author{
I. Ramos-Gutiérrez, R. Gumbs
}
