% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE2_multiphylo}
\alias{calculate_EDGE2_multiphylo}
\title{EDGE2 calculating function for a multiPhylo tree object.}
\usage{
calculate_EDGE2_multiphylo(
  multitree,
  table,
  species.col = "species",
  RLcat.col = "RLcat",
  sort.list = FALSE,
  ext.prob = "Isaac",
  verbose = TRUE,
  return.all = FALSE,
  seed = NULL,
  summarise = TRUE,
  parallelize = FALSE,
  n.cores = NULL,
  ...
)
}
\arguments{
\item{multitree}{a `multiPhylo` object storing multiple trees}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{species.col}{column storing species names. If not specified, a column named "species" will be searched for.}

\item{RLcat.col}{column storing Red List assessments names. If not specified, a column named "RLcat" will be searched for.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{ext.prob}{Extinction probability associated to category, based on Isaac et al. (2007) or Mooers et al. (2008).
Accepted values are 'Isaac', 'IUCN50', 'IUCN100' and 'IUCN500'}

\item{verbose}{Logical. Should progress be printed or not.}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}

\item{summarise}{Logical. If FALSE, a list of dataframes outputted from `calculate_EDGE1` function will be returned,
one for each tree stored in the 'multitree' `multiPhylo` object. If FALSE, results will be summarised into a single table,
with mean, standard deviation, median and inter-quartile ranges for ED and EDGE metrics across trees.}

\item{parallelize}{Logical. If TRUE, several CPU cores (defined in `n.cores`) will be used.}

\item{n.cores}{Integer. Number of cores to use simultaneoulsly. Default is the number of available ones minus one.}
}
\value{
A list of length equal to the number of trees included in the 'multiPhylo' object, storing information resulting
from the `calculate_EDGE2` function used based on the specified method. If `return.all` is set to TRUE, the returned object will be a list
storing individual results of the `calculate_EDGE2` without summarising and including also `tree` and `ePDloss` for each tree in the 'multitree' object.
}
\description{
EDGE2 calculating function for a multiPhylo tree object.
}
\author{
I. Ramos-Gutiérrez.
}
