% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE_multiple.R
\name{calculate_EDGE2_multiple}
\alias{calculate_EDGE2_multiple}
\title{Multiple iteration EDGE2 calculating function.}
\usage{
calculate_EDGE2_multiple(
  tree,
  table,
  species.col = "species",
  RLcat.col = "RLcat",
  sort.list = FALSE,
  ext.prob = "Isaac",
  return.all = FALSE,
  summarise = TRUE,
  n.iter = 10,
  parallelize = FALSE,
  n.cores = NULL,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{tree}{A 'phylo' or 'multiPhylo' object.}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{species.col}{column storing species names. If not specified, a column named "species" will be searched for.}

\item{RLcat.col}{column storing Red List assessments names. If not specified, a column named "RLcat" will be searched for.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{ext.prob}{Extinction probability associated to category, based on Isaac et al. (2007) or Mooers et al. (2008).
Accepted values are 'Isaac', 'IUCN50', 'IUCN100' and 'IUCN500'}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}

\item{summarise}{Logical. If FALSE, a list of dataframes outputted from `calculate_EDGE1` function will be returned,
one for each tree stored in the 'multitree' `multiPhylo` object. If FALSE, results will be summarised into a single table,
with mean, standard deviation, median and inter-quartile ranges for ED and EDGE metrics across trees.}

\item{n.iter}{Integer. Number of times the function will be run.}

\item{parallelize}{Logical. If TRUE, several CPU cores (defined in `n.cores`) will be used.}

\item{n.cores}{Integer. Number of cores to use simultaneoulsly. Default is the number of available ones minus one.}

\item{verbose}{Logical. Should progress be printed or not.}
}
\value{
A list of length = n.iter
}
\description{
Multiple iteration EDGE2 calculating function.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
calculate_EDGE2_multiple(monotreme.tree,
                        monotreme.table,
                        verbose = T,
                        sort.list = T,
                        n.iter = 4,
                        parallelize = T,
                        n.cores = 4,
                        seed = 123456)
# example code
\dontshow{\}) # examplesIf}
}
\author{
I. Ramos-Gutiérrez.
}
