% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE_wrappers.R
\name{calculate_EDGE_multiphylo}
\alias{calculate_EDGE_multiphylo}
\title{EDGE2 calculating function for a multiPhylo tree object.}
\usage{
calculate_EDGE_multiphylo(
  multiphylo,
  table,
  method = "EDGE2",
  ext.prob = "Isaac",
  sort.list = FALSE,
  return.all = FALSE,
  summarise = TRUE,
  parallelize = FALSE,
  n.cores = NULL,
  seed = NULL
)
}
\arguments{
\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}

\item{ext.prob}{Extinction probability associated to category, based on Isaac et al. (2007) or Mooers et al. (2008).}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}

\item{summarise}{Logical. IF TRUE (default), median and IQR values across iterations will be summarise. Only applicable when return.all is FALSE.}

\item{parallelize}{Logical. If TRUE, several CPU cores will be used to compute EDGE scores.}

\item{n.cores}{Integer. Number of cores to use simultaneoulsly. Default is the number of available ones minus one.}

\item{seed}{Integer. seed number used to randomize iterations. if NULL, a random number will be enerated and printed}
}
\value{
A list of length equal to the number of trees included in the multiphylo object, storing
information resulting from the calculate_EDGE function used based on the specified method.
}
\description{
EDGE2 calculating function for a multiPhylo tree object.
}
\author{
I. Ramos-Gutiérrez.
}
