% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDi.R
\name{calculate_PD_indicator}
\alias{calculate_PD_indicator}
\title{Phylogenetic Diversity indicator calculating function}
\usage{
calculate_PD_indicator(multiphylo, table, time.cols, ...)
}
\arguments{
\item{multiphylo}{An object of class mutliPhylo where to calculate the PDi}

\item{table}{a data.frame containing a column names 'species' with all taxa in the tree,
and columns storing Red List categories.}

\item{time.cols}{column names of all times whose PDi should be calculated.}
}
\value{
a list including a dataframe dor each time column, where PD, ePDloss and PDi are shown
}
\description{
Phylogenetic Diversity indicator calculating function
}
\examples{
 calculate_PD_indicator(multiphylo = crocodile.trees,
                        table = crocodile.table,
                        time.cols = c("RL.2023", "RL.2024"))


}
\author{
I. Ramos-Gutiérrez, R. Gumbs
}
