% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crocodile.table}
\alias{crocodile.table}
\title{crocodile.table}
\format{
An info data frame
\describe{
Data frame with Red List assessments for the 27 species of monotremes for years 2023 & 2024.

}
}
\usage{
crocodile.table
}
\description{
Information data frame with Red List assessments for the species in the crocodile phylogeny for years 2023 and 2024.
}
\keyword{datasets}
