% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crocodile.trees}
\alias{crocodile.trees}
\title{crocodile.trees}
\format{
A multiPhylo object storing 10 crocodile phylogenies
\describe{
multiPhylo object with 10 phylogenies including 27 species of crocodiles.
}
}
\usage{
crocodile.trees
}
\description{
Example phylogeny for rEDGE package
}
\keyword{datasets}
