% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycad.table}
\alias{cycad.table}
\title{cycad.table}
\format{
An data frame
\describe{
Data frame with Red List assessments for the 337 species of cycads for years 2003 & 2014.

}
}
\usage{
cycad.table
}
\description{
Information data frame with Red List assessments for cycad species in two time frames, namely 2003 and 2014.
}
\keyword{datasets}
