% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycad.tree}
\alias{cycad.tree}
\title{cycad.tree}
\format{
A phylo object
\describe{
Phylo object including 337 species of cycads.
}
}
\usage{
cycad.tree
}
\description{
Phylogeny of cycads oftained from Forest et al. (2018)
}
\keyword{datasets}
