% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monotreme.table}
\alias{monotreme.table}
\title{monotreme.table}
\format{
An info data frame
\describe{
Data frame with Red List assessments for the 5 species of monotremes (IUCN, September 2025)

}
}
\usage{
monotreme.table
}
\description{
Information data frame with Red List assessments for the species in the monotreme phylogeny.
}
\keyword{datasets}
