% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monotreme.tree}
\alias{monotreme.tree}
\title{Phylogeny of monotremes}
\format{
A phylogenetic tree with 5 tips, including platypus and echidnas.
\describe{
Phylogenetic tree with 5 species of monotremes (Monotremata).
This is a subtree extracted from the complete mammal phylogeny by Upham et al. (2019)
}
}
\usage{
monotreme.tree
}
\description{
Example phylogeny for EDGEcalc package
}
\keyword{datasets}
