% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE_wrappers.R
\name{calculate_EDGE}
\alias{calculate_EDGE}
\title{Calculate EDGE wrapping function}
\usage{
calculate_EDGE(tree, table, method = "EDGE2", n.iter = NULL, ...)
}
\arguments{
\item{tree}{Phylogenetic information as a 'phylo' or 'multiPhylo' object}

\item{table}{a tibble or data frame with two columns, including species names and Red List categories.}

\item{method}{Method to be used. It has to be either "EDGE1" or "EDGE2" (default).}
}
\description{
Calculate EDGE wrapping function
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

# Calculate EDGE-1 for a single tree
calculate_EDGE(monotreme.tree,
monotreme.table,
method = "EDGE1")

# Calculate EDGE-1 for a multiPhylo object (multiple phylogenies)
calculate_EDGE(cycad.multitree, cycad.table,
               method = "EDGE1",
               RLcat.col = "RL_2003", # in this example it is not the default name (RLcat)!
               sort.list = T)

# Other variations
calculate_EDGE(cycad.multitree,
               cycad.table,
               method = "EDGE1",
               RLcat.col = "RL_2014",
               sort.list = FALSE,
               summarise = FALSE,
               parallelize = TRUE,
               n.cores = 4)


# Calculate EDGE-2 for a single tree
calculate_EDGE(monotreme.tree,
               monotreme.table,
               method = "EDGE2",
               seed=123)

calculate_EDGE(monotreme.tree,
               monotreme.table,
               method = "EDGE2",
               sort.list = FALSE,
               ext.prob = "IUCN100",
               seed=123)

# Calculate EDGE-2 for a multiPhylo object (multiple phylogenies)
calculate_EDGE(cycad.multitree,
               cycad.table,
               method = "EDGE2",
               RLcat.col = "RL_2014",
               sort.list = FALSE,
               summarise = FALSE,
               parallelize = TRUE,
               n.cores = 4)

# Calculate EDGE-2 for a single tree, but 100 times
calculate_EDGE(monotreme.tree,
               monotreme.table,
               n.iter = 100)

# Calculate EDGE-2 for a multiple phylogenies, and 10 times each, combining other parameters
calculate_EDGE(tree = cycad.multitree,
               table = cycad.table,
               RLcat.col = "RL_2003",
               ext.prob ="IUCN50",
               n.iter = 10,
               parallelize = T,
               sort.list = T,
               return.all = F,
               summarise = F)
\dontshow{\}) # examplesIf}
}
