% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE}
\alias{calculate_EDGE}
\title{Calculate EDGE wrapping function}
\usage{
calculate_EDGE(tree, table, method = "EDGE2", ...)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}

\item{method}{Method to be used. It has to be either "EDGE1" or "EDGE2".
The output will be the result of running calculate_EDGE1, or calculate_EDGE2 depending on the method specified.}
}
\description{
Calculate EDGE wrapping function
}
