% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE1.R
\name{calculate_EDGE1}
\alias{calculate_EDGE1}
\title{Calculate EDGE index}
\usage{
calculate_EDGE1(tree, table)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}
}
\value{
a data frame with the input table and the EDGE metric (sensu  Isaac et al., 2007).
}
\description{
Calculate EDGE index
}
