% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE2}
\alias{calculate_EDGE2}
\title{EDGE2 calculating function.}
\usage{
calculate_EDGE2(tree, table, verbose = T, sort.list = FALSE)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}

\item{verbose}{Logical. Should progress be printed or not.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}
}
\value{
A list containing 3 slots.
\itemize{
 \item{list: }{will output specific TBL (terminal branch length), pext (sampled probability of extinction), ED (evolutionary distinctiveness), and EDGE (EDGE index).}
 \item{tree: }{the input tree.}
 \item{ePDloss: }{complete PD (phylogenetic diversity) and ePDloss (expected PD loss). Units are as in the input tree  (generally Million years).}
}
}
\description{
EDGE2 calculating function.
}
\author{
I. Ramos-Gutiérrez, R. Gumbs
}
