% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE2}
\alias{calculate_EDGE2}
\title{EDGE2 calculating function.}
\usage{
calculate_EDGE2(
  tree,
  table,
  verbose = TRUE,
  sort.list = FALSE,
  return.all = FALSE
)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}

\item{verbose}{Logical. Should progress be printed or not.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}
}
\value{
A list containing 3 slots if `return.all` is TRUE. If not, only the first item (EDGE list) will be returned.
\itemize{
 \item{list: }{will output specific TBL (terminal branch length), pext (sampled probability of extinction), ED (evolutionary distinctiveness), and EDGE (EDGE index).}
 \item{tree: }{the input tree.}
 \item{ePDloss: }{complete PD (phylogenetic diversity) and ePDloss (expected PD loss). Units are as in the input tree  (generally Million years).}
}
}
\description{
EDGE2 calculating function.
}
\references{
Gumbs, R., Gray, C. L., Böhm, M., Burfield, I. J., Couchman, O. R., Faith, D. P.,
Forest, F., Hoffmann, M., #' Isaac, N. J. B., Jetz, W., Mace, G. M., Mooers, A. O., Safi, K.,
Scott, O., Steel, M., Tucker, C. M., Pearse, W. D., Owen, N. R. & Rosindell, J. (2023).
The EDGE2 protocol: Advancing the prioritisation of Evolutionarily Distinct and Globally Endangered species for
practical conservation action. PLoS Biology, 21(2), e3001991.
\url{https://doi.org/10.1371/journal.pbio.3001991}
}
\author{
I. Ramos-Gutiérrez, R. Gumbs
}
