% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDGE2.R
\name{calculate_EDGE2_multiple}
\alias{calculate_EDGE2_multiple}
\title{Multiple iteration EDGE2 calculating function.}
\usage{
calculate_EDGE2_multiple(
  tree,
  table,
  verbose = T,
  sort.list = FALSE,
  return.all = FALSE,
  n.iter = 100,
  parallelize = FALSE,
  n.cores = NULL
)
}
\arguments{
\item{tree}{a phylo object.}

\item{table}{a tibble or data frame with two columns, named "species" and "RL.cat"}

\item{verbose}{Logical. Should progress be printed or not.}

\item{sort.list}{Logical. If TRUE, the EDGE list will be sorted from higher to lower values.}

\item{return.all}{Logical. If TRUE, an EDGE list, tree and ePDloss list is returned. If FALSE (default), only the list is returned.}

\item{n.iter}{Integer. Number of times the function will be run.}

\item{parallelize}{Logical. If TRUE, several CPU cores will be used to compute EDGE scores.}

\item{n.cores}{Integer. Number of cores to use simultaneoulsly. Default is the number of available ones minus one.}
}
\value{
A list of length = n.iter
}
\description{
Multiple iteration EDGE2 calculating function.
}
\author{
I. Ramos-Gutiérrez.
}
