% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycad.multitree}
\alias{cycad.multitree}
\title{cycad.multitree}
\format{
A mutliPhylo object
\describe{
100 phylogenies object including 337 species of cycads.
}
}
\usage{
cycad.multitree
}
\description{
100 phylogenies of cycads oftained from Forest et al. (2018)
}
\keyword{datasets}
